test cases Need to figure out dependency injection for the BitStream classes. Problem is the methods I need to make fail are templates, so I can't just make them virtual and supply a BitStream subclass for testing.

+++ fly/coders/huffman/huffman_encoder.cpp
@@ -35,7 +35,7 @@ bool HuffmanEncoder::EncodeInternal(
     {
         LOGW(
             "Maximum Huffman code length %u is too large for code_type",
-            m_maxCodeLength);
+            static_cast<std::uint32_t>(m_maxCodeLength));
         return false;
     }
     else if (!encodeHeader(output))
